+++ build.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -euo pipefail
 
 #export DEBUG=true
 
@@ -13,7 +14,7 @@ popd
 
 # Release Build of the Docker Image
 # Final Docker Image Name
-export IMAGE_BUILD="r0"
+export IMAGE_BUILD="r1"
 export DOCKER_REPO=konvergence/duplicity
 
 
@@ -31,11 +32,15 @@ export DOCKER_REPO=konvergence/duplicity
 #	read -p "release folder:" RELEASE_FOLDER
 #fi
 
+export PG_VERSION=""
+echo "Building filesystem image"
+./buildDockerImage.sh
+
+
 PG_VERSIONS="12 14 15"
 
-#${ALL_RELEASES_FOLDER}/${RELEASE_FOLDER}/buildDockerImage.sh
 for PG_VERSION in ${PG_VERSIONS}; do
     export PG_VERSION
-    echo "Building image with PG_VERSION=$PG_VERSION"
+    echo "Building postgresql image with PG_VERSION=$PG_VERSION"
     ./buildDockerImage.sh
 done